bpf2go: export compilation
Move the code necessary to compile a C to an ELF. The behaviour of
bpf2go is unchanged.

The code to fix up make-style depfiles remains in bpf2go since it
has little tests, and is probably used only seldomly.
