+++ cmd/bpf2go/gen/doc.go
@@ -1,0 +1,2 @@
+// Package gen contains utilities to generate Go bindings for eBPF ELF files.
+package gen
